+++ freight/__init__.py
@@ -1,4 +1,4 @@
 default_app_config = 'freight.apps.FreightConfig'
 
-__version__ = '1.1.0a1'
+__version__ = '1.1.0a2'
 __title__ = 'Alliance Freight'